tests/requirements-test.txt
# Test dependencies for Content Verification Tool

# Core testing framework
pytest>=8.0.0
pytest-asyncio>=0.23.0

# FastAPI testing
httpx>=0.25.0

# Coverage reporting (optional)
pytest-cov>=4.1.0

# Parallel test execution (optional)
pytest-xdist>=3.5.0

# Test output formatting
termcolor>=2.3.0

# Document generation testing
pandas>=2.1.0
openpyxl>=3.1.0
python-docx>=1.1.0

# Google Gemini SDK (for AI tests)
google-genai>=1.0.0

# Environment variables
python-dotenv>=1.0.0
